requirements.txt
anthropic
neris_api_client
python-dotenv
